Prime number
A number is defined as Prime number . when it consists of only two factors. to count the number of factors we took a variable called facount. whenever the number is divided by the another number we keep on increasing the facount of 1. and finally after the all iterations if we get facount is 2 then it is a prime number. otherwise it is not a prime number.

Signed-off-by: Devarapalli vasanthi <Vasanthidevarapalli17@gmail.com>